rpc 接口优化

+++ src/main/java/com/dx/facade/venue/dto/DxTableInfoDto.java
@@ -31,7 +31,7 @@ public class DxTableInfoDto implements Serializable {
     private String tableName;
 
     @ApiModelProperty("牌桌类型 2001：德州局 2002：短牌局")
-    private Integer tableType;
+    private Long tableType;
 
     @ApiModelProperty("关联游戏id")
     private Long gameTypeId;
